+++ src/Nemiro.OAuth/Clients/TwitterClient.cs
@@ -1,5 +1,5 @@
 ﻿// ----------------------------------------------------------------------------
-// Copyright © Aleksey Nemiro, 2014-2015. All rights reserved.
+// Copyright © Aleksey Nemiro, 2014-2016. All rights reserved.
 // 
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -414,6 +414,7 @@ namespace Nemiro.OAuth.Clients
       map.Add("profile_image_url", "Userpic");
       map.Add("url", "Url");
       map.Add("birthday", "Birthday", typeof(DateTime), @"dd\.MM\.yyyy");
+      map.Add("lang", "Language");
       //map.Add("verified", "Url");
       //map.Add("location", "Url");