+++ src/Nemiro.OAuth/Clients/YandexClient.cs
@@ -1,5 +1,5 @@
 ﻿// ----------------------------------------------------------------------------
-// Copyright © Aleksey Nemiro, 2014-2015. All rights reserved.
+// Copyright © Aleksey Nemiro, 2014-2016. All rights reserved.
 // 
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -419,12 +419,16 @@ namespace Nemiro.OAuth.Clients
       // query parameters
       var parameters = new NameValueCollection
       { 
-        { "format", "json" },
-        { "oauth_token", accessToken.Value },
+        { "format", "json" }
       };
 
       // execute the request
-      var result = OAuthUtility.Get("https://login.yandex.ru/info", parameters);
+      var result = OAuthUtility.Get
+      (
+        "https://login.yandex.ru/info", 
+        parameters: parameters, 
+        authorization: new HttpAuthorization(AuthorizationType.OAuth, accessToken.Value)
+      );
 
       // field mapping
       var map = new ApiDataMapping();